Assorted Refit Drone Friend fixes (#217)
* Don't spawn befriended refit drone with isAiming trait as true

With upcoming UITR status lights, the overwatch state was visible until
the drone moved.

* Prevent PetDrone from appearing on the other side of a wall

canUnitReach was happy to show a BLOCKED message in that case.

Likewise for RenameDrone, but don't show messages when unavailable
to avoid clutter.

Fixes #213

* Preview the AP boost when hovering the PetDrone button

to go with UITR's push towards AP-modifying abilities using the vanilla
previewAbilityAP functions as appropriate.

* Incorporate Makefile improvements from UITR

Only builds and copies files as needed

* Add commentary about cover and guard hostility in Drone Friend tooltip

* Prevent RefitDroneFriend from pathing through enemy lethal lasers

Thanks Sizzle for figuring this out first for Dragon.

* Tweak wording: Remove conjunction between statements

* Suppress the AP Boost preview until after the player has seen the effect

Using the seenOnce settings map, normally used for controlling modals like
"has seen the rewind tutorial."

* Use UITR's helper for AP-boost/cost ability tooltips, with fallback

If a recent UITR version isn't present, falls back to just showing the
tooltip, same as it used to before.

* Update Makefile to use oil_fx.kwad instead of pedler_oil.kwad

At some point, the kwad was renamed and the modinit load updated to the
new name. Bringing the Makefile up to match.

* Ensure that the settings seenOnce table is initialized before use

`viz_manager.lua` does this, so we should as well.